add ui file and rename the gui files to be more descriptive
