+++ varats/varats/gui/cs_gen/case_study_generation_ui.py
